+++ syntheseus/tests/interface/test_models.py
@@ -2,7 +2,7 @@ import math
 
 import numpy as np
 
-from syntheseus.interface.models import Prediction, PredictionList
+from syntheseus.interface.models import Prediction
 from syntheseus.interface.molecule import Molecule
 
 # TODO(kmaziarz): Currently this test mostly checks that importing from `models.py` works, and that
@@ -14,10 +14,4 @@ def test_prediction():
     assert np.isclose(prediction.get_prob(), 0.5)
     assert np.isclose(prediction.get_log_prob(), math.log(0.5))
 
-    other_prediction = Prediction(input=Molecule("N"), output=Molecule("NC=O"), probability=0.5)
-    prediction_list = PredictionList(
-        input=Molecule("C"), predictions=[prediction, other_prediction]
-    )
-
-    assert prediction_list.predictions == [prediction, other_prediction]
-    assert prediction_list.truncated(num_results=1).predictions == [prediction]
+    Prediction(input=Molecule("N"), output=Molecule("NC=O"), probability=0.5)